+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Clients require implementations of `RequestFactoryInterface`
 - Renamed `Notion\Client::createWithPsrClient()` to `createWithPsrImplementations()`
+- Renamed `Notion\Client` to `Notion\Notion`
 
 ## [0.1.0] - 2021-11-03
 ### Added